fixed minor errors
git-svn-id: svn+ssh://voms.jku.austriangrid.at/var/subversion/proc@47 593b3eb8-86d1-0410-841a-be20d09665e6

+++ jkutop.h
@@ -55,8 +55,8 @@ typedef struct pidstat
 	int					euid;
 	int					pid;
 	int					ppid;
-	unsigned long		majflt;
-	unsigned long		minflt;
+	unsigned long long	majflt;
+	unsigned long long	minflt;
 	unsigned long long	utime;
 	unsigned long long	stime;
 	unsigned long long	utime_lastpass;
